requirements.txt
numpy
matplotlib
scikit-image
opencv-python-headless
Pillow
pytesseract
scipy
os
cv2